Converted all Int64 to Int32 Created APi for Chnage Password and Get Login Details

+++ 365_Portal/Code/BL/AchivementBL.cs
@@ -15,7 +15,7 @@ namespace _365_Portal.Code.BL
             Logger.Log(ex, "AchivementBL", methodName);
         }
 
-        public static DataSet CreateAchievement(Int64 CompId, Int64 AchievementId,  string AchivementTitle, string AchivementDescription, string LongDescription, string CreatedBy)
+        public static DataSet CreateAchievement(Int32 CompId, Int32 AchievementId,  string AchivementTitle, string AchivementDescription, string LongDescription, string CreatedBy)
         {
             DataSet ds = new DataSet();
             try
@@ -29,7 +29,7 @@ namespace _365_Portal.Code.BL
             return ds;
         }
 
-        public static DataSet DeleteAchievement(Int64 CompId, Int64 AchievementId)
+        public static DataSet DeleteAchievement(Int32 CompId, Int32 AchievementId)
         {
             DataSet ds = new DataSet();
             try
@@ -43,7 +43,7 @@ namespace _365_Portal.Code.BL
             return ds;
         }
 
-        public static DataSet GetAchievements(Int64 CompId, Int64 AchievementId)
+        public static DataSet GetAchievements(Int32 CompId, Int32 AchievementId)
         {
             DataSet ds = new DataSet();
             try
