+++ 365_Portal/Code/DAL/FlashcardDAL.cs
@@ -15,7 +15,7 @@ namespace _365_Portal.Code.DAL
             Logger.Log(ex, "FlashcardDAL", methodName);
         }
 
-        public static DataSet CreateFlashcard(Int64 CompId, Int64 ContentId, Int64 FlashcardId, string FlashcardTitle, string FlashcardDescription, Int64 SrNo, string CreatedBy)
+        public static DataSet CreateFlashcard(Int32 CompId, Int32 ContentId, Int32 FlashcardId, string FlashcardTitle, string FlashcardDescription, Int32 SrNo, string CreatedBy)
         {
 
             DataSet ds = new DataSet();
@@ -51,7 +51,7 @@ namespace _365_Portal.Code.DAL
             return ds;
         }
 
-        public static DataSet DeleteFlashcard(Int64 CompId, Int64 ContentId, string CreatedBy)
+        public static DataSet DeleteFlashcard(Int32 CompId, Int32 ContentId, string CreatedBy)
         {
 
             DataSet ds = new DataSet();
@@ -83,7 +83,7 @@ namespace _365_Portal.Code.DAL
             return ds;
         }
 
-        public static DataSet GetFlashcards(Int64 CompId, Int64 ContentId)
+        public static DataSet GetFlashcards(Int32 CompId, Int32 ContentId)
         {
 
             DataSet ds = new DataSet();
